removed azhRunLocalCmd from all.js
